Use MCCos, not MCSin, for the server predicted motion's Z

+++ check/basic.go
@@ -11,7 +11,7 @@ type basic struct {
 
 // Buff adds to the buffer and returns the new one.
 func (t *basic) Buff(n float64, max ...float64) float64 {
-	m := float64(15)
+	m := 15.0
 	if len(max) > 0 {
 		m = max[0]
 	}
